requirements.txt
cryptography==44.0.2
Django==5.1.6
django-stubs-ext==5.2.2
langchain==0.3.19
langchain-chroma==0.2.2
langchain-community==0.3.18
langchain-openai==0.3.7
mysqlclient==2.2.7
PyMuPDF==1.25.3
PyMySQL==1.1.1
pyTelegramBotAPI==4.26.0